requirements.txt
annotated-types==0.7.0
app==0.0.1
app-model==0.5.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.2
in-n-out==0.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.2
psygnal==0.14.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
gspread==5.12.4
google-auth-oauthlib==1.2.0
